requirements.txt.txt
streamlit==1.28.0
pytesseract==0.3.10
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
requests==2.31.0